Removing demo files, removing duplicate interfaces from merge, moving article and articles pages to the pages folder
