+++ src/app/app.module.ts
@@ -8,11 +8,9 @@ import { AppComponent } from './app.component';
 import { ArticleTypeMainComponent } from './components/articles/article-type-main/article-type-main.component';
 import { ArticleTypeSideComponent } from './components/articles/article-type-side/article-type-side.component';
 import { ArticleTypeSmallComponent } from './components/articles/article-type-small/article-type-small.component';
-import { ArticlesComponent } from './components/articles/articles.component';
-import { DemoComponent } from './components/demo/demo.component';
-import { DemoDirective } from './directives/demo.directive';
+import { ArticlesComponent } from './pages/articles/articles.component';
 
-import { ArticleComponent } from './components/articles/article/article.component';
+import { ArticleComponent } from './pages/article/article.component';
 import { TitleComponent } from './components/articles/article/title/title.component';
 import { BodyComponent } from './components/articles/article/body/body.component';
 import { OtherComponent } from './components/articles/article/other/other.component';
@@ -20,12 +18,8 @@ import { OtherComponent } from './components/articles/article/other/other.compon
 import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
 
 
-import { DemoPageComponent } from './pages/demo/demo.component';
-import { DemoPipe } from './pipes/demo.pipe';
-
-
 @NgModule({
-  declarations: [AppComponent, DemoPageComponent, DemoComponent, DemoPipe, DemoDirective, ArticlesComponent, ArticleTypeMainComponent, ArticleTypeSideComponent, ArticleTypeSmallComponent, PageNotFoundComponent, ArticleComponent, TitleComponent, BodyComponent, OtherComponent],
+  declarations: [AppComponent, ArticlesComponent, ArticleTypeMainComponent, ArticleTypeSideComponent, ArticleTypeSmallComponent, PageNotFoundComponent, ArticleComponent, TitleComponent, BodyComponent, OtherComponent],
   imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule],
   providers: [],
   bootstrap: [AppComponent],
