--- src/app/category.ts
@@ -1,4 +1,0 @@
-export interface Category {
-    id: number
-    name: string
-}
