--- src/app/components/demo/demo.component.html
@@ -1,1 +1,0 @@
-<h1 appDemo>{{ title | demo }}</h1>
