--- src/app/components/demo/demo.component.scss
@@ -1,4 +1,0 @@
-h1 {
-  padding: 1rem;
-  margin: 0;
-}
