--- src/app/components/demo/demo.component.ts
@@ -1,11 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-demo-component',
-  templateUrl: './demo.component.html',
-  styleUrls: ['./demo.component.scss'],
-})
-export class DemoComponent {
-  @Input()
-  title = '';
-}
