+++ src/app/pages/article/article.component.ts
