--- src/app/pages/demo/demo.component.html
@@ -1,1 +1,0 @@
-
