--- src/app/pages/demo/demo.component.scss
@@ -1,7 +1,0 @@
-main {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 0.5rem;
-}
