--- src/app/services/demo.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class DemoService {
-  findAllLinter() {
-    return ['ESlint', 'Prettier'];
-  }
-}
