--- src/app/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-    id: number
-    firstname: string
-    lastname: string
-}
